requirements.txt
pandas
python-dotenv
google-generativeai
pydantic
streamlit
sqlalchemy
